# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "type": "object"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "params",
          "requestId"
        ],
        "type": "object"
      },
...
      "Tenant": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "role"
        ],
        "type": "object"
      },
      "TenantCreated": {
        "additionalProperties": false,
        "properties": {
          "space": {
            "$ref": "#/components/schemas/Space"
          },
          "tenant": {
            "$ref": "#/components/schemas/Tenant"
          }
        },
        "required": [
          "tenant",
          "space"
        ],
        "type": "object"
      },
...
      "User": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "displayName",
          "status",
          "version",
          "createdAt",
          "deletedAt"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "serviceCredential": {
        "bearerFormat": "EdDSA JWT",
        "description": "Pinned issuer/kid/kind=service/role, aud=ora-cloud, exp and iat required, \u003c=5 minute lifetime. Public API requires gateway; internal control requires controller; nodes require scoped node role.",
        "scheme": "bearer",
        "type": "http"
      },
      "userCredential": {
        "description": "Separately signed EdDSA JWT: kind=user, source+sub, caller must equal authenticated service sub, aud=ora-cloud. User and membership status checked in PostgreSQL.",
        "in": "header",
        "name": "X-Ora-User-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Authoritative PostgreSQL core. Simulation is separate; no production Controller/Node/Kubernetes implementation is implied.",
    "title": "Ora Cloud phase one",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/me": {
      "get": {
        "description": "Public requests require a gateway service credential plus a caller-bound user credential. Tenant membership is checked before lookup; resource reads filter tenant and owner in SQL. Mutation version conflicts return 409; a missing required version returns 428. Unknown fields are rejected. Lists use ascending UUID pagination.",
        "operationId": "get_api_v1_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful command or resource response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid JSON/field/input, missing idempotency key, invalid pagination or evidence"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid, forged, expired, wrong-audience, untrusted, or caller-mismatched credential"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Disabled user, inactive/missing membership, wrong service role, or admin required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource absent or outside authorized tenant/owner scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Version/idempotency conflict, resource_in_use, closed admission, stale epoch/Node/sandbox, incomplete effect, invalid transition, unconfirmed termination/idle, last_admin, space_last_owner, or space_slug_conflict"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Version precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal error; no SQL or secret details are exposed"
          }
        },
        "security": [
          {
            "serviceCredential": [],
            "userCredential": []
          }
        ],
        "summary": "GET /api/v1/me",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/tenants": {
      "get": {
        "description": "Public requests require a gateway service credential plus a caller-bound user credential. Tenant membership is checked before lookup; resource reads filter tenant and owner in SQL. Mutation version conflicts return 409; a missing required version returns 428. Unknown fields are rejected. Lists page in ascending creation order; the after cursor is an exclusive tenant UUID.",
        "operationId": "get_api_v1_me_tenants",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Exclusive UUID cursor, ascending stable ordering.",
            "in": "query",
            "name": "after",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/Tenant"
                      },
                      "type": "array"
                    },
                    "nextCursor": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful command or resource response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid JSON/field/input, missing idempotency key, invalid pagination or evidence"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid, forged, expired, wrong-audience, untrusted, or caller-mismatched credential"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Disabled user, inactive/missing membership, wrong service role, or admin required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource absent or outside authorized tenant/owner scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Version/idempotency conflict, resource_in_use, closed admission, stale epoch/Node/sandbox, incomplete effect, invalid transition, unconfirmed termination/idle, last_admin, space_last_owner, or space_slug_conflict"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Version precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal error; no SQL or secret details are exposed"
          }
        },
        "security": [
          {
            "serviceCredential": [],
            "userCredential": []
          }
        ],
        "summary": "GET /api/v1/me/tenants",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/tenants": {
      "post": {
        "description": "Public requests require a gateway service credential plus a caller-bound user credential. No tenant membership is required: the verified identity alone authorizes provisioning. Atomically creates a tenant named after the space, makes the caller its first administrator, creates the space with the given slug and makes the caller its owner. The tenant is an implicit container the product never shows. The idempotency key is matched per user across tenants and recorded under the created tenant. Mutation version conflicts return 409; a missing required version returns 428. Unknown fields are rejected. Lists use ascending UUID pagination.",
        "operationId": "post_api_v1_tenants",
        "parameters": [
          {
            "description": "Scoped to tenant and user. Same key and canonical method/path/body returns the original response before version validation; changed request is 409.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "slug": {
                    "description": "Lowercase, immutable, unique per tenant.",
                    "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "slug"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantCreated"
                }
              }
            },
            "description": "Successful command or resource response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid JSON/field/input, missing idempotency key, invalid pagination or evidence"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid, forged, expired, wrong-audience, untrusted, or caller-mismatched credential"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Disabled user, inactive/missing membership, wrong service role, or admin required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource absent or outside authorized tenant/owner scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Version/idempotency conflict, resource_in_use, closed admission, stale epoch/Node/sandbox, incomplete effect, invalid transition, unconfirmed termination/idle, last_admin, space_last_owner, or space_slug_conflict"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Version precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal error; no SQL or secret details are exposed"
          }
        },
        "security": [
          {
            "serviceCredential": [],
            "userCredential": []
          }
        ],
        "summary": "POST /api/v1/tenants",
        "tags": [
          "tenants"
        ]
      }
    },
...
    "/api/v1/tenants/{tid}/clones/{cid}": {
      "get": {
        "description": "Public requests require a gateway service credential plus a caller-bound user credential. Tenant membership is checked before lookup; resource reads filter tenant and owner in SQL. Clone requests are independent accepted work items outside the project/workspace operation model: Cloud accepts them in its own transaction, a Controller claims and dispatches them over the internal control contract, and only the submitting user can read them. requestId is the caller's durable request identity: repeating it with the same repository and branch returns the original request, a different input is 409 idempotency_conflict. repository must be an https or ssh URL the Controller can clone; branch is a short branch name, never HEAD. executionId and nodeId are null until a dispatch is recorded; a pending state means awaiting reconciliation, never failure. Mutation version conflicts return 409; a missing required version returns 428. Unknown fields are rejected. Lists use ascending UUID pagination.",
        "operationId": "get_api_v1_tenants_tid_clones_cid",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloneOperation"
                }
              }
            },
            "description": "Successful command or resource response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid JSON/field/input, missing idempotency key, invalid pagination or evidence"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid, forged, expired, wrong-audience, untrusted, or caller-mismatched credential"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Disabled user, inactive/missing membership, wrong service role, or admin required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource absent or outside authorized tenant/owner scope"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Version/idempotency conflict, resource_in_use, closed admission, stale epoch/Node/sandbox, incomplete effect, invalid transition, unconfirmed termination/idle, last_admin, space_last_owner, or space_slug_conflict"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Version precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal error; no SQL or secret details are exposed"
          }
        },
        "security": [
          {
            "serviceCredential": [],
            "userCredential": []
          }
        ],
        "summary": "GET /api/v1/tenants/:tid/clones/:cid",
        "tags": [
          "clones"
        ]
      }
    },
...
    "/api/v1/tenants/{tid}/spaces/{sid}/events": {
      "get": {
        "description": "Membership is verified before the stream opens. Events are lightweight invalidation notices published after commit; clients refetch authoritative state over REST.",
        "operationId": "getSpaceEvents",
        "parameters": [
          {
            "in": "path",
            "name": "tid",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sid",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceEvent"
                }
              }
            },
            "description": "Server-sent event stream"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid, forged, expired, wrong-audience, untrusted, or caller-mismatched credential"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Disabled user, inactive/missing membership, wrong service role, or admin required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Resource absent or outside authorized tenant/owner scope"
          }
        },
        "security": [
          {
            "serviceCredential": [],
            "userCredential": []
          }
        ],
        "summary": "Stream workspace events over server-sent events",
        "tags": [
          "spaces"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ]
}